+++ src/main/java/com/api/managepersonaddresses/repositories/PersonRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.api.managepersonaddresses.models.PersonModel;
 
 
-public interface PersonRepository extends JpaRepository<PersonModel, Integer> {
+public interface PersonRepository extends JpaRepository<PersonModel, Long> {
   
   boolean existsByName(String name);
 
-  Optional<PersonModel> findById(Integer id);
+  Optional<PersonModel> findById(Long id);
 
 }
